Split to StudyHistory component
